Now using Invector to control the character
Removed our old garbage way of moving the character, and now the character's movements and animations are handled by Invector. To add an animation: Go to mixamo, select your animation, check "In Place" if the option is present, then hit Download. The Format should be FBX for Unity, FPS should be 30, keyframe reduction should be none, and Skin should be "without skin". Download and drop the animation into Unity, in the Assets/Characters/Man Dude/Animations folder. Select the new animation, and in the inspector, change Rig > Animation Type to Humanoid, then Materials > Location to Use External Materials (Legacy). Apply the changes, then you can duplicate the animation itself out from the folder with @ in its name, and delete the @ folder.