Started adding the fishing minigame
Added the fishing minigame scene. Added a new item and model for a fishing rod. Also added a new custom 3d model for a lake, and animated low-poly water from the asset store. Actions now check to see if conditions are met before performing the animation, such as checking if the player can carry an item before attempting pickup. Added the fishing interact action, which requires the player be able to carry the new fish, as well as having a fishing rod in the inventory.