Started adding the Item Details UI
The basic UI has been added for the Item Details popup. A new script has been created to control the Item Details objects. Clicking an item in the inventory will now open the item details screen, which features a rotating model of the item. Also packaged the Canvas, Event Listener, and Cameras into a new "Essentials" prefab for ease of use in developing future minigame levels.