Mega UI changes
Added the MainUI script and associated UI elements, which includes an action button, an (unusable) attack button, a pause button, a placeholder for a joystick, and placeholders for Health and Sanity status info.

The player inventory has been moved to the new PauseUI script and UI elements, which is accessible via the Pause button. The inventory has been resized, repositioned, and rescaled, and the number of inventory slots has changed, but still functions the same as before. Going to the pause menu also pauses time for the minigame.

Item quantity has been added to the item details UI.

Moved the Town UI to its own scene.