Woodcutting game - Deposit/Finish Functions
Added new logic to the Deposit Handler to take in wood, and deliver the quantity of wood collected at the end of the game. Updated the essentials prefab with the deposit handler. Updated the inventory to have some more functionality like "Clear Inventory" I also made the script check to see if the inventory that's being edited is the player inventory to prevent null reference errors.