Added a new player character model
Added the Explorer Female rig and model, along with the rest of the Polygon Explorer Kit. Created the Explorer Female prefab, which now works as a substitute for the Man Dude character in a scene.

Changed the unity script execution order so that the Essentials script has its start method run before almost all other scripts, which guarantees that the StaticVariables values are defined before other scripts need to reference them.