Dynamic Player Model Instantiation
Added the PlayerModelSetup script to the essentials prefab, which contains a list of possible player model options, as well as a basic method to select which player model should be used in the current scene. Man Dude has been replaced with the Player Spawn Point in all scenes. When the scene starts, the PlayerModelOptions script removes the Player Spawn Point and replaces it with an instance of the chosen player model object.