Improved the Player Model Setup script
Player Model Setup has an int field to input the index of the player model you want to use, and another int field to input the index of the material you want to apply to that model.

At the start of the scene, the player model is instantiated, the chosen material is applied, and then the Player Model Setup object gets deleted from the scene.