Implemented Silas' UI and added Character Stats
The Pause Menu and Character Creation Menus have been updated to use Silas' new UI elements.

Added a CharacterStats script, which contains basic information - name, inventory size, health, and sanity, as well as the character's vehicle. Alongside the vehicle are the transform properties that should be applied to the vehicle so it appears in the Character Selection screen properly.

Changing characters in the Character Selection screen now updates the Character Stats panel.